+++ offline-2/src/HeuristicData.hpp
@@ -5,6 +5,6 @@
 class HeuristicData
 {
 public:
-    static uint8_t heuristic1, heuristic2;
+    static uint16_t heuristic1, heuristic2;
     static uint64_t w11, w12, w13, w14, w21, w22, w23, w24;
 };
